requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.1
matplotlib==3.8.3
jinja2==3.1.3
pyyaml==6.0.1
pytz==2024.1
python-dotenv==1.0.1
